fixed resizing of already existing windows.
Clients for these windows were created right before the app dispatches
`on_ready`. Because of that, the `on_new_client` event handler of
sx-simpledeco wasn't installed at this time. Because of that, the
sx-simpledeco plugin couldn't set an actor before `Client.init`
registered the actor for the `on_configure_notify` window. Because of
that, the client listened to configure notify events of the window and
had the wrong geometry set.
Now, all screens are scanned right after dispatching `on_ready`. That
should be better, I hope there are no side effects in existing plugins.