samurai-x2: added _NET_WM_VISIBLE_NAME, _NET_WM_NAME and WM_NAME support
The more important change is that `image_text16` works with Python
unicode objects now. The key is the 'utf-16be' encoding, which creates
a string that matches CHAR2B structures.
I added a switch in `ooxcb_client.py` for that. It could be that
QueryTextExtends & co won't work for now because of that change.